feat(schema-utils): new package; extract Standard Schema introspection from validate (TP-017)
Stands up `@crustjs/schema-utils` 0.0.1, a new published workspace package
containing the vendor-aware Standard Schema introspection layer that
previously lived inside `@crustjs/validate`.

Unblocks TP-018 (`@crustjs/store` consumes the same introspection without
depending on `@crustjs/validate`).

## @crustjs/schema-utils (new, 0.0.1)

Pre-stability release; surface unstable until 0.1.0. Four helper groups:

| Group | Exports |
|-------|---------|
| Introspection | `inferOptions`, `extractDefault`, `InferredOptions`, `ExtractedDefault` |
| Boundary | `isStandardSchema`, `assertStandardSchema` |
| Issues | `normalizeStandardIssues`, `normalizeStandardPath`, `formatPath`, `ValidationIssue` |
| Types | `StandardSchema`, `InferInput`, `InferOutput` |

Vendor-aware dispatch on `~standard.vendor` \u2014 Zod and Effect today; other
vendors (Valibot, ArkType, etc.) fall through to `{}` per the existing
contract.

## @crustjs/validate (no observable API change)

The TP-014 locked root surface (`arg`, `flag`, `commandValidator`,
`field`, `parseValue`, `validateStandard`, `validateStandardSync`,
`isStandardSchema`, plus type re-exports) is unchanged. Introspection
physically moves to schema-utils; validate re-exports transparently.
Consumers do not need to update any imports.

Adds one transitive dependency (`@crustjs/schema-utils`). No behavior
change \u2014 every call site continues to receive identical inferred metadata,
identical default extraction, identical thrown error shapes
(`CrustError("DEFINITION")`).

## File moves

Renames preserved by git (R099 = 99% similarity):

- `packages/validate/src/introspect/registry.ts` \u2192 `packages/schema-utils/src/introspect.ts`
- `packages/validate/src/introspect/registry.test.ts` \u2192 `packages/schema-utils/src/introspect.test.ts`
- `packages/validate/src/introspect/zod.ts` \u2192 `packages/schema-utils/src/zod.ts`
- `packages/validate/src/introspect/effect.ts` \u2192 `packages/schema-utils/src/effect.ts`

New files extract assertion/issue helpers from validate:
- `packages/schema-utils/src/assertions.ts` (from `validate/src/validate.ts`)
- `packages/schema-utils/src/issues.ts` (from `validate/src/validate.ts`)
- `packages/schema-utils/src/types.ts` (re-export aliases)
- `packages/schema-utils/src/index.ts` (barrel)

## Verification

- `bun run check` \u2705 287 files, no fixes
- `bun run check:types --force` \u2705 23/23 (was 21 \u2192 +2 for new package)
- `packages/schema-utils` tests: 23 pass / 0 fail / 39 expect calls
- `packages/validate` tests: 293 pass / 0 fail / 509 expect calls (was
  316/548 \u2014 the 23 tests / 39 expects moved to schema-utils, byte-identical)
- All 11 packages: 0 fail
