+++ README.md
@@ -47,6 +47,7 @@ bun run dev
 | [`@crustjs/progress`](https://www.npmjs.com/package/@crustjs/progress) | Progress indicators for async CLI tasks                                          | [![npm](https://img.shields.io/npm/v/@crustjs/progress)](https://www.npmjs.com/package/@crustjs/progress) | [![downloads](https://img.shields.io/npm/dm/@crustjs/progress)](https://www.npmjs.com/package/@crustjs/progress) |
 | [`@crustjs/prompts`](https://www.npmjs.com/package/@crustjs/prompts)   | Interactive terminal prompts                                                     | [![npm](https://img.shields.io/npm/v/@crustjs/prompts)](https://www.npmjs.com/package/@crustjs/prompts)   | [![downloads](https://img.shields.io/npm/dm/@crustjs/prompts)](https://www.npmjs.com/package/@crustjs/prompts)   |
 | [`@crustjs/validate`](https://www.npmjs.com/package/@crustjs/validate) | Validation helpers                                                               | [![npm](https://img.shields.io/npm/v/@crustjs/validate)](https://www.npmjs.com/package/@crustjs/validate) | [![downloads](https://img.shields.io/npm/dm/@crustjs/validate)](https://www.npmjs.com/package/@crustjs/validate) |
+| [`@crustjs/schema-utils`](https://www.npmjs.com/package/@crustjs/schema-utils) | Vendor-aware Standard Schema introspection helpers                          | [![npm](https://img.shields.io/npm/v/@crustjs/schema-utils)](https://www.npmjs.com/package/@crustjs/schema-utils) | [![downloads](https://img.shields.io/npm/dm/@crustjs/schema-utils)](https://www.npmjs.com/package/@crustjs/schema-utils) |
 | [`@crustjs/store`](https://www.npmjs.com/package/@crustjs/store)       | DX-first, typed persistence for CLI apps with config/data/state/cache separation | [![npm](https://img.shields.io/npm/v/@crustjs/store)](https://www.npmjs.com/package/@crustjs/store)       | [![downloads](https://img.shields.io/npm/dm/@crustjs/store)](https://www.npmjs.com/package/@crustjs/store)       |
 | [`@crustjs/skills`](https://www.npmjs.com/package/@crustjs/skills)     | Agent skill generation from Crust command definitions                            | [![npm](https://img.shields.io/npm/v/@crustjs/skills)](https://www.npmjs.com/package/@crustjs/skills)     | [![downloads](https://img.shields.io/npm/dm/@crustjs/skills)](https://www.npmjs.com/package/@crustjs/skills)     |
 | [`@crustjs/create`](https://www.npmjs.com/package/@crustjs/create)     | Headless scaffolding engine for building create-xxx tools                        | [![npm](https://img.shields.io/npm/v/@crustjs/create)](https://www.npmjs.com/package/@crustjs/create)     | [![downloads](https://img.shields.io/npm/dm/@crustjs/create)](https://www.npmjs.com/package/@crustjs/create)     |
