+++ apps/docs/content/docs/modules/meta.json
@@ -7,6 +7,7 @@
     "crust",
     "create",
     "validate",
+    "schema-utils",
     "progress",
     "prompts",
     "style",
