+++ apps/docs/content/docs/modules/schema-utils.mdx
@@ -1,0 +1,124 @@
+---
+title: "Schema Utils"
+description: Vendor-aware Standard Schema introspection helpers shared by @crustjs/validate and @crustjs/store.
+---
+
+import { Callout } from "fumadocs-ui/components/callout";
+
+`@crustjs/schema-utils` exposes the **vendor-aware introspection layer** that
+`@crustjs/validate` and `@crustjs/store` use to recover metadata from any
+[Standard Schema v1](https://standardschema.dev/) object. Given a schema, it
+dispatches on `~standard.vendor` to return the `type`, `array` shape,
+`description`, `default`, and normalized issue paths that a CLI or a
+persistent store needs.
+
+<Callout type="warn">
+**Pre-stability — version `0.0.1`.** Most users do not need to install this
+package directly. The high-level packages (`@crustjs/validate`,
+`@crustjs/store`) consume it internally. Only depend on it when authoring
+framework-level tooling on top of Standard Schema. Pin an exact version —
+the surface is unstable until `0.1.0`.
+</Callout>
+
+## Install
+
+```sh
+bun add @crustjs/schema-utils
+```
+
+## Helpers at a glance
+
+| Group        | Symbols                                                | Use when…                                              |
+| ------------ | ------------------------------------------------------ | ------------------------------------------------------ |
+| `introspect` | `inferOptions`, `extractDefault`                       | You need `type` / `array` / `description` / `default`. |
+| `assertions` | `isStandardSchema`, `assertStandardSchema`             | You need a boundary check on untrusted input.          |
+| `issues`     | `normalizeStandardIssues`, `normalizeStandardPath`     | You're surfacing validation errors to a user.          |
+| `types`      | `StandardSchema`, `InferInput`, `InferOutput`          | You need a portable type alias.                        |
+
+## Quick example — `inferOptions`
+
+```ts title="introspect-zod.ts"
+import { inferOptions } from "@crustjs/schema-utils";
+import { z } from "zod";
+
+const schema = z.string().describe("hello");
+const opts = inferOptions(schema, "field", "label");
+// → { type: "string", multiple: false, description: "hello", optional: false }
+```
+
+`inferOptions` accepts a `StandardSchema` plus a `kind` (`"arg"` / `"flag"` /
+`"field"`) and a label used in vendor-specific error messages. It returns a
+partial `InferredOptions` object — fields are absent when the underlying
+vendor cannot tell us. For unknown vendors it returns `{}` (the consumer
+falls back to user-supplied options).
+
+## Quick example — `extractDefault`
+
+```ts title="extract-default.ts"
+import { extractDefault } from "@crustjs/schema-utils";
+import { z } from "zod";
+
+extractDefault(z.number().default(3000));
+// → { ok: true, value: 3000 }
+
+extractDefault(z.string());
+// → { ok: false }
+```
+
+`extractDefault` is **synchronous** by design — store-field initialization
+needs the value before any `await` is allowed. It runs a vendor-specific
+walker first (Zod's `ZodDefault`, Effect's `Default` annotation), then falls
+back to a vendor-neutral `validate(undefined)` probe.
+
+<Callout>
+Effect schemas annotated only via `Schema.annotations({ default })` cannot
+inject defaults during decode and therefore won't be extracted. Use
+`Schema.optionalWith(Schema, { default })` (or `.default()` in Zod) to opt
+in to default-extraction.
+</Callout>
+
+## Boundary assertion
+
+Use `assertStandardSchema(value, label)` at API boundaries that accept
+arbitrary input and want to fail fast on the wrong type:
+
+```ts
+import { assertStandardSchema } from "@crustjs/schema-utils";
+
+function field(schema: unknown) {
+	assertStandardSchema(schema, "field");
+	// schema is now typed as StandardSchema
+}
+```
+
+`isStandardSchema(value)` is the matching boolean type guard.
+
+## Issue normalization
+
+```ts
+import { normalizeStandardIssues } from "@crustjs/schema-utils";
+
+const result = await schema["~standard"].validate(input);
+if (result.issues) {
+	const issues = normalizeStandardIssues(result.issues, ["flags", "verbose"]);
+	// → [{ message: "...", path: "flags.verbose" }]
+}
+```
+
+`normalizeStandardIssues(issues, prefix?)` flattens Standard Schema issue
+paths to a dot-path string, applying an optional path prefix that the caller
+supplies (e.g. `["flags", "verbose"]` so the final path reads
+`flags.verbose`).
+
+## Types
+
+`StandardSchema<Input, Output>` is a thin alias over `StandardSchemaV1`.
+`InferInput<S>` and `InferOutput<S>` give you the input/output types of a
+schema instance. Re-exported by `@crustjs/validate` so most consumers never
+import from this package directly.
+
+## Primary consumers
+
+- [`@crustjs/validate`](/docs/modules/validate) — CLI argument, flag, and
+  store-field helpers built on top of `inferOptions` / `extractDefault`.
+- `@crustjs/store` — store-field metadata extraction (TP-018).
