+++ packages/schema-utils/bunup.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from "bunup";
+
+export default defineConfig({
+	entry: ["src/index.ts"],
+	format: ["esm"],
+	target: "bun",
+	dts: true,
+	minify: true,
+});
