+++ packages/schema-utils/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "@crustjs/config/tsconfig.base.json",
+	"compilerOptions": {
+		"declaration": true,
+		"isolatedDeclarations": true
+	},
+	"include": ["src", "tests"]
+}
